Over-the-air Updates for fortunaxx (20180501)
Over-the-air updates for both supra and stable branch users for the day of 1st May, 2018 (20180501), dated 23rd April, 2018.

Changelog:
* The May One Security Patch
* MxPlatform API version 4.3.0
* Streamlining of system applications, fixing system hang
* Kernel upstreaming, updates, and optimizations
* Performance tweaks and improvements
* Fixes to font rendering and anti-aliasing

Updates to the following applications:
* ClockPackage - MX4.3_2018A
* SamsungCalendar_MXP - MX4.3
* SamsungIME - v5_SYSTEMUID_SWIFTKEY_VO
* Messaging - MX4.3_2018A
* SamsungVideoPlayer - 2018A_MX4.3
* SecGallery - 2018A
* SecMyFiles - 2018A_MX4.3
* SystemUI - 2018A_MX4.3

NOTE: This system update is the same on both supra and stable branches, as this was approved to be released passively (on all branches) after closed beta testing.